refactor: move preset command handlers to commands/preset.py
Extracts all preset_app and preset_catalog_app command handlers out of
the monolithic __init__.py into src/specify_cli/commands/preset.py.
Updates test patch target for _locate_bundled_preset to the new module path.
